+++ packages/matter.js/test/behavior/cluster/ClusterBehaviorTest.ts
@@ -29,10 +29,27 @@ import { My, MyBehavior, MyCluster } from "./cluster-behavior-test-util.js";
 class MyEventedBehavior extends MyBehavior {
     declare events: MyEventedBehavior.Events;
 }
+
 namespace MyEventedBehavior {
     export class Events extends MyBehavior.Events {
         somethingHappened = new Observable<[]>();
     }
+}
+
+class BehaviorWithCustomMethods extends MyBehavior.with("Awesome") {
+    foo() {
+        return true;
+    }
+
+    bar() {
+        return 4;
+    }
+}
+
+namespace BehaviorWithCustomMethods {
+    export declare const ExtensionInterface: {
+        bar(): number;
+    };
 }
 
 describe("ClusterBehavior", () => {
@@ -170,6 +187,29 @@ describe("ClusterBehavior", () => {
             expect(MyBehavior.defaults.attr2).equals(123);
             expect(MyBehavior.defaults.attr3).equals("abc");
             expect(MyBehavior.defaults.attr4).equals(undefined);
+        });
+
+        it("carries forward non-command methods", () => {
+            ({}) as BehaviorWithCustomMethods satisfies { foo(): boolean };
+
+            const WithRevertedFeature = BehaviorWithCustomMethods.for(MyCluster);
+
+            ({}) as InstanceType<typeof WithRevertedFeature> satisfies { foo(): boolean };
+
+            class FooOverrideBehavior extends WithRevertedFeature {
+                // Note -- limitation due to https://github.com/microsoft/TypeScript/issues/27965 requires us to
+                // override as an instance function rather than a method
+                override foo = () => {
+                    return false;
+                };
+
+                // This is in ExtensionInterface so TS bug above does not apply
+                override bar() {
+                    return 5;
+                }
+            }
+
+            ({}) as InstanceType<typeof FooOverrideBehavior> satisfies { foo(): boolean };
         });
     });